moved tests from algebras file to collections file to mirror the function files
