+++ test/collections.spec.js
@@ -33,6 +33,126 @@ describe('Collections Functions', () => {
     expect(F.findApply('a', xs)).to.equal(1)
     expect(F.findApply('d', xs)).to.equal(undefined)
   })
+  it('map', () => {
+    // map plain arrays
+    expect(F.map(x => x * x, [1, 2, 3])).to.deep.equal([1, 4, 9])
+    // map plain objects
+    expect(F.map(x => x * x, { a: 1, b: 2, c: 3 })).to.deep.equal({
+      a: 1,
+      b: 4,
+      c: 9,
+    })
+  })
+  describe('deepMap', () => {
+    it('deepMap arrays', () => {
+      // arrays
+      const arr = [0, [1, [2, []]]]
+      const arrBackup = _.cloneDeep(arr)
+      const arrMutated = F.deepMap(e => e.concat(101), arr)
+      //      Checking immutability
+      expect(arr).to.eql(arrBackup)
+      expect(arrMutated).to.eql([0, [1, [2, [101], 101], 101]])
+    })
+    it('deepMap plain objects', () => {
+      const objA = {
+        a: {
+          match: {
+            id: 1,
+          },
+          b: {
+            match: {
+              id: 2,
+            },
+            c: {
+              match: {
+                id: 3,
+              },
+            },
+          },
+        },
+      }
+      const objABackup = _.cloneDeep(objA)
+      const pathA = 'match.matched'
+      const setMatchedA = e => e.match && _.set(pathA, true, e)
+      const objAMutated = F.deepMap(e => setMatchedA(e) || e)(objA)
+      //      Checking immutability
+      expect(objA).to.eql(objABackup)
+      expect(objAMutated).to.eql({
+        a: {
+          match: {
+            id: 1,
+            matched: true,
+          },
+          b: {
+            match: {
+              id: 2,
+              matched: true,
+            },
+            c: {
+              match: {
+                id: 3,
+                matched: true,
+              },
+            },
+          },
+        },
+      })
+    })
+    it('deepMap with plain objects with arrays with objects', () => {
+      const objB = {
+        a: {
+          array: [0, [1, [2, [{ match: { id: 0 } }]]]],
+          match: {
+            id: 1,
+          },
+          b: {
+            match: {
+              id: 2,
+            },
+            c: {
+              match: {
+                id: 3,
+              },
+            },
+          },
+        },
+      }
+      const objBBackup = _.cloneDeep(objB)
+      const pathB = 'match.matched'
+      const setMatchedB = e => e.match && _.set(pathB, true, e)
+      const push101 = e => _.isArray(e) && e.concat(101)
+      const objBMutated = F.deepMap(e => push101(e) || setMatchedB(e) || e)(
+        objB
+      )
+      //         Checking immutability
+      expect(objB).to.eql(objBBackup)
+      expect(objBMutated).to.eql({
+        a: {
+          array: [
+            0,
+            [1, [2, [{ match: { id: 0, matched: true } }, 101], 101], 101],
+            101,
+          ],
+          match: {
+            id: 1,
+            matched: true,
+          },
+          b: {
+            match: {
+              id: 2,
+              matched: true,
+            },
+            c: {
+              match: {
+                id: 3,
+                matched: true,
+              },
+            },
+          },
+        },
+      })
+    })
+  })
   it('insertAtIndex', () => {
     let arr = [1, 2, 3]
     let x = F.insertAtIndex(1, 5, arr)
